requirements.txt
quart               # Основной фреймворк
bcrypt              # Для хэширования
python-dotenv       # Для переменных окружения
asyncpg             # Для асинхронной PostgreSQL
